pyproject.toml
[tool.poetry]
name = "strafrecht"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "4.9.0"
bleach = "6.1.0"                    # Remove candidate
django-avatar = "8.0.1"
django-colorfield = "0.4.1"         # Remove candidate
django-comments-xtd = "2.9.13"
django-contrib-comments = "2.2.0"
django-cors-headers = "3.7.0"       # Remove candidate. Shoud be implemented on ALB level
django-crispy-forms = "1.14.0"
django-debug-toolbar = "3.2.1"
django-filter = "23.5"
django-ipware = "1.1.2"             # Remove candidate, used once
django-mjml = "1.3"
django-modelcluster = "6.3"
django-nyt = "1.4.1"
django-pagedown = "2.2.0"           # Remove candidate, used once
django-pipeline = "3.1.0"           # Remove candidate, should be a spearate Frontend process
django-pwa = "2.0.1"
django-sekizai = "2.0.0"
django-taggit = "4.0.0"
django-extensions = "^3.2.3"
django = "4.2.16"                   # Need update as it's not maintained now
djangorestframework = "3.14.0"
lxml = "5.3.0"
markdownify = "0.13.1"              # Remove candidate, used once
psycopg2-binary = "2.9.10"
python-dateutil = "2.9.0"           # Remove candidate, used once in a simple case
python-magic = "0.4.27"             # Remove candidate, used once
requests = "2.32.3"
sorl-thumbnail = "12.11.0"
wagtail-autocomplete = "0.11.0"
wagtail-birdsong = "2.0.0"          # Remove candidate, used once
wagtail-color-panel = "1.5.0"
wagtail-modelchooser = "4.0.1"
wagtail-treemodeladmin = "1.9.2"
wagtail = "5.2.6"
wagtailmenus = "4.0.1"
wagtailstreamforms = "4.1.0"        # Remove candidate, blocks upgrading to WagTail 6.3
wget = "3.2"                        # Remove candidate, used once, can be replaced with requests
wiki = "0.10.1"                     # Remove candidate, duplicate of CMS feature. Has lots of secondary deps
wrapt = "1.14.1"                    # Remove candidate, used once, in a simple case

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
